Trying to fix movement code
Works for one target, have hit wall when it comes to multiple targets. Got a line of (1px x 1px) rectangle objects spawning as a trail but it is too taxing for the system and I have yet to work out the ID system which would allow removal of a line associated with a particular warhead.